+++ CHANGELOG.md
@@ -4,10 +4,9 @@ All notable changes to ag-grid-date-filter will be documented in this file.
 
 ## [1.1.0](https://github.com/ryanrozich/ag-grid-react-components/compare/v1.1.0-ag-grid-bug-identified...v1.1.0) (2025-06-24)
 
-
 ### ✨ Features
 
-* **utils:** add AG Grid v33 setFilterModel workaround and documentation ([73b3034](https://github.com/ryanrozich/ag-grid-react-components/commit/73b3034f7cfb277981680af82fab35bc168b591c)), closes [ag-grid/ag-grid#2256](https://github.com/ag-grid/ag-grid/issues/2256) [ag-grid/ag-grid#2709](https://github.com/ag-grid/ag-grid/issues/2709) [ag-grid/ag-grid#4870](https://github.com/ag-grid/ag-grid/issues/4870)
+- **utils:** add AG Grid v33 setFilterModel workaround and documentation ([73b3034](https://github.com/ryanrozich/ag-grid-react-components/commit/73b3034f7cfb277981680af82fab35bc168b591c)), closes [ag-grid/ag-grid#2256](https://github.com/ag-grid/ag-grid/issues/2256) [ag-grid/ag-grid#2709](https://github.com/ag-grid/ag-grid/issues/2709) [ag-grid/ag-grid#4870](https://github.com/ag-grid/ag-grid/issues/4870)
 
 ## [1.0.0] - 2025-06-08